move fixtures to their own files
