+++ tests_fp/cross_platform/fixtures/project.py
@@ -1,0 +1,13 @@
+"""Project fixture"""
+from pathlib import Path
+
+import ape
+from ape.managers.project import ProjectManager
+import pytest
+
+
+@pytest.fixture(scope="function")
+def project() -> ProjectManager:
+    "Returns the ape project."
+    project_root = Path.cwd()
+    return ape.Project(path=project_root)
