+++ tests_fp/cross_platform/utils.py
@@ -6,9 +6,6 @@ As such, we are relaxing some of the lint rules.
 """
 from __future__ import annotations
 
-# stdlib
-from collections import defaultdict
-
 # external lib
 import ape
 from ape.contracts import ContractInstance
@@ -17,7 +14,8 @@ from eth_typing import HexAddress
 # elfpy core repo
 import elfpy.markets.hyperdrive.hyperdrive_assets as hyperdrive_assets
 import elfpy.markets.hyperdrive.hyperdrive_market as hyperdrive_market
-from tests_fp.cross_platform.conftest import HyperdriveConfig
+from elfpy.math.fixed_point import FixedPoint
+from tests_fp.cross_platform.fixtures.hyperdrive_config import HyperdriveConfig
 
 
 def to_fixed_point(float_var: float, decimal_places=18):
@@ -34,11 +32,11 @@ def get_simulation_market_state_from_contract(
     hyperdrive_contract: ContractInstance,
     hyperdrive_data_contract: ContractInstance,
     agent_address: HexAddress,
-    position_duration_seconds: int,
-    checkpoint_duration_days: int,
-    variable_apr: int,
+    position_duration_seconds: FixedPoint,
+    checkpoint_duration_days: FixedPoint,
+    variable_apr: FixedPoint,
     config: HyperdriveConfig,
-) -> hyperdrive_market.MarketState:
+) -> hyperdrive_market.HyperdriveMarketState:
     """
     hyperdrive_contract: ape.contracts.base.ContractInstance
         Ape project `ContractInstance
@@ -51,37 +49,37 @@ def get_simulation_market_state_from_contract(
     """
     # pylint: disable=too-many-arguments
     pool_info = hyperdrive_data_contract.getPoolInfo()
-    # pool_state = pool_info.__dict__
-    pool_state = defaultdict(lambda: 0)
+    pool_state = pool_info.__dict__
+    # pool_state = defaultdict(lambda: 0)
     with ape.accounts.use_sender(agent_address):  # sender for contract calls
         asset_id = hyperdrive_assets.encode_asset_id(
-            hyperdrive_assets.AssetIdPrefix.WITHDRAWAL_SHARE, position_duration_seconds
+            hyperdrive_assets.AssetIdPrefix.WITHDRAWAL_SHARE, int(int(position_duration_seconds) / 10**18)
         )
         total_supply_withdraw_shares = hyperdrive_data_contract.balanceOf(asset_id, agent_address)
 
-    return hyperdrive_market.MarketState(
-        lp_total_supply=to_floating_point(pool_state["lpTotalSupply"]),
-        share_reserves=to_floating_point(pool_state["shareReserves"]),
-        bond_reserves=to_floating_point(pool_state["bondReserves"]),
-        base_buffer=pool_state["longsOutstanding"],  # so do we not need any buffers now?
+    return hyperdrive_market.HyperdriveMarketState(
+        lp_total_supply=FixedPoint(scaled_value=pool_state["lpTotalSupply"]),
+        share_reserves=FixedPoint(scaled_value=pool_state["shareReserves"]),
+        bond_reserves=FixedPoint(scaled_value=pool_state["bondReserves"]),
+        base_buffer=FixedPoint(scaled_value=pool_state["longsOutstanding"]),  # so do we not need any buffers now?
         # TODO: bond_buffer=0,
         variable_apr=variable_apr,
-        share_price=to_floating_point(pool_state["sharePrice"]),
+        share_price=FixedPoint(scaled_value=pool_state["sharePrice"]),
         init_share_price=config.share_price,
         curve_fee_multiple=config.curve_fee,
         flat_fee_multiple=config.flat_fee,
-        longs_outstanding=to_floating_point(pool_state["longsOutstanding"]),
-        shorts_outstanding=to_floating_point(pool_state["shortsOutstanding"]),
-        long_average_maturity_time=to_floating_point(pool_state["longAverageMaturityTime"]),
-        short_average_maturity_time=to_floating_point(pool_state["shortAverageMaturityTime"]),
-        long_base_volume=0,  # to_floating_point(pool_state["longBaseVolume"]),
-        short_base_volume=to_floating_point(pool_state["shortBaseVolume"]),
+        longs_outstanding=FixedPoint(scaled_value=pool_state["longsOutstanding"]),
+        shorts_outstanding=FixedPoint(scaled_value=pool_state["shortsOutstanding"]),
+        long_average_maturity_time=FixedPoint(scaled_value=pool_state["longAverageMaturityTime"]),
+        short_average_maturity_time=FixedPoint(scaled_value=pool_state["shortAverageMaturityTime"]),
+        long_base_volume=FixedPoint(0),  # FixedPoint(scaled_value=pool_state["longBaseVolume"]),
+        short_base_volume=FixedPoint(scaled_value=pool_state["shortBaseVolume"]),
         # TODO: checkpoints=defaultdict
         checkpoint_duration=checkpoint_duration_days,
-        total_supply_longs=defaultdict(float, {0: to_floating_point(pool_state["longsOutstanding"])}),
-        total_supply_shorts=defaultdict(float, {0: to_floating_point(pool_state["shortsOutstanding"])}),
-        total_supply_withdraw_shares=to_floating_point(total_supply_withdraw_shares),
-        withdraw_shares_ready_to_withdraw=to_floating_point(pool_state["withdrawalSharesReadyToWithdraw"]),
-        withdraw_capital=to_floating_point(pool_state["withdrawalSharesProceeds"]),
-        withdraw_interest=0,  # to_floating_point(pool_state["interest"]),
+        total_supply_longs={FixedPoint(0): FixedPoint(scaled_value=pool_state["longsOutstanding"])},
+        total_supply_shorts={FixedPoint(0): FixedPoint(scaled_value=pool_state["shortsOutstanding"])},
+        total_supply_withdraw_shares=FixedPoint(scaled_value=total_supply_withdraw_shares),
+        withdraw_shares_ready_to_withdraw=FixedPoint(scaled_value=pool_state["withdrawalSharesReadyToWithdraw"]),
+        withdraw_capital=FixedPoint(scaled_value=pool_state["withdrawalSharesProceeds"]),
+        withdraw_interest=FixedPoint(0),  # FixedPoint(scaled_value=pool_state["interest"]),
     )
